Admin code is not much more manageable
